+++ src/Mapping/DistinctValueSet.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace GlpiPlugin\Ticketmigration\Mapping;
+
+final readonly class DistinctValueSet
+{
+    public function __construct(public array $values, public bool $truncated) {}
+}